fix: address sleep widget review findings
- Use targeted Dexie query (entryForDate) instead of loading all entries
- Pass paraglide locale to toLocaleTimeString for consistent formatting
- Add loading state to prevent flash of empty-state CTA
- Suppress "0m" when sleep duration is exact hours
- Add "See more" header link to sleep insights tab
- Add .superpowers/ to .gitignore

+++ src/lib/server/validation/responses/analytics.ts
@@ -55,8 +55,8 @@ const sleepFoodCorrelationEntrySchema = z
 	.object({
 		date: z.string(),
 		eveningCalories: z.number().nullable(),
-		sleepDurationMinutes: z.number(),
-		sleepQuality: z.number()
+		sleepDurationMinutes: z.number().int(),
+		sleepQuality: z.number().int()
 	})
 	.meta({ id: 'SleepFoodCorrelationEntry' });
